+++ plugin-parent/file-system/src/main/java/jscover/maven/FileTestRunner.java
@@ -45,6 +45,10 @@ public class FileTestRunner {
         if (jsonFile.exists())
             jsonFile.delete();
         if (config.isLocalStorage()) {
+            //TODO Can remove below 2 lines when upgrading to JSCover 1.0.12
+            String resource = "jscoverage-clear-local-storage.html";
+            ioUtils.copy(getClass().getResourceAsStream("/" + resource), new File(config.getDestDir(), resource));
+            webClient.get("file:///" + new File(config.getDestDir(), resource).getAbsolutePath().replaceAll("\\\\","/"));
             for (File testPage : testPages)
                 runTestLocalStorage(ioUtils.getRelativePath(testPage, config.getDestDir()));
             String json = (String)((JavascriptExecutor) webClient).executeScript("return jscoverage_serializeCoverageToJSON();");