+++ test.sh
@@ -1,0 +1,1 @@
+mvn clean cobertura:cobertura -Dcobertura.aggregate=true && mvn surefire-report:report -DskipTests=true -Daggregate=true